requirements.txt
mkdocs>=1.5.0
mkdocs-material>=9.0.0
pyyaml>=6.0.0
mkdocs-minify-plugin>=0.6.0
